Fix compiler warning

+++ Color Extensions/UIColor+HTMLColors.m
@@ -374,13 +374,13 @@ static NSCharacterSet *CMRHexCharacters() {
 // We know we've got hex already, so assume this works
 static NSUInteger CMRParseHex(NSString *str, BOOL repeated)
 {
-    NSUInteger ans = 0;
+    unsigned int ans = 0;
     if (repeated) {
         str = [NSString stringWithFormat:@"%@%@", str, str];
     }
     NSScanner *scanner = [NSScanner scannerWithString:str];
     [scanner scanHexInt:&ans];
-    return ans;
+    return (NSUInteger)ans;
 }
 
 // Scan FFF or FFFFFF, doesn't reset scan location on failure
